% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_codebook_creator.R
\name{codebook_server}
\alias{codebook_server}
\title{Codebook Generator Server Module}
\usage{
codebook_server(id)
}
\arguments{
\item{id}{A unique module identifier.}
}
\description{
Codebook Generator Server Module
}
