% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_codebook_creator.R
\name{codebookServer}
\alias{codebookServer}
\title{Codebook Generator Server Module}
\usage{
codebookServer(id)
}
\arguments{
\item{id}{A unique module identifier.}
}
\description{
Codebook Generator Server Module
}
