% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_codebook_creator.R
\name{codebookUI}
\alias{codebookUI}
\title{Codebook Generator Module}
\usage{
codebookUI(id)
}
\arguments{
\item{id}{A unique module identifier.}
}
\value{
A Shiny module for uploading a CSV, viewing/editing a codebook, and downloading the result.
}
\description{
This module provides a Shiny UI and server logic to generate a codebook from a CSV file.
}
