% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataframe.R
\name{compare_data_tags}
\alias{compare_data_tags}
\title{Compare Metadata Between Two Data Objects}
\usage{
compare_data_tags(df1, df2)
}
\arguments{
\item{df1}{First data object with attributes.}

\item{df2}{Second data object with attributes.}
}
\value{
A tibble with comparison results or NA if attributes missing.
}
\description{
Compares key metadata attributes between two datasets.
}
