% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blank_pipeline.R
\name{create_blank_pipeline}
\alias{create_blank_pipeline}
\title{Create a Blank Data Analysis Pipeline}
\usage{
create_blank_pipeline(
  root_path = getwd(),
  pipeline_folder_name = "demo-pipeline",
  additional_folders = NULL,
  verbose = TRUE
)
}
\arguments{
\item{root_path}{Character. The root path where the pipeline folder should be created. Default is the current working directory.}

\item{pipeline_folder_name}{Character. The name of the pipeline folder to be created. Default is "demo-pipeline".}

\item{additional_folders}{Character vector. Optional additional subfolders to create inside the pipeline.}

\item{verbose}{Logical. If TRUE, prints messages indicating folder and file creation. Default is TRUE.}
}
\value{
None. The function creates folders and files in the specified location.
}
\description{
This function initializes a structured project for a data analysis pipeline,
including essential folders (`data`, `plots`, `tidy`, `vignettes`) and README files.
}
\examples{
\dontrun{
create_blank_pipeline("~/Desktop", "my_pipeline", additional_folders = c("scripts", "models"))
}
}
