% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variables.R
\name{extract_variables_from_file}
\alias{extract_variables_from_file}
\title{Extract Column Names from an Excel File}
\usage{
extract_variables_from_file(file)
}
\arguments{
\item{file}{A string representing the file path to the Excel file.}
}
\value{
A tibble with two columns: `dataframe` (file name) and `column_name` (column names in the file).
}
\description{
This function reads an Excel file using `tidypipes::read_data_file` and
extracts its column names, returning a tibble with the file name and column names.
}
