% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variables.R
\name{extract_variables_from_file_list}
\alias{extract_variables_from_file_list}
\title{Extract Column Names from a List of Excel Files}
\usage{
extract_variables_from_file_list(files)
}
\arguments{
\item{files}{A list of strings, each representing the file path to an Excel file.}
}
\value{
A tibble with two columns: `filename` (file name) and `column_name` (column names in the files).
}
\description{
This function applies `extract_variables_from_file` to each file in a list
and combines the results into a single tibble.
}
