% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_info.R
\name{get_column_info}
\alias{get_column_info}
\title{Get Column Names and Data Types}
\usage{
get_column_info(dataset)
}
\arguments{
\item{dataset}{A data frame or tibble for which the column names and data types are to be extracted.}
}
\value{
A tibble containing the column names and their corresponding data types.
}
\description{
This function generates a tibble with the column names and data types for all columns in a given dataset.
}
\examples{
# Create a sample dataset
dataset <- tibble::tibble(
  col1 = 1:5,
  col2 = letters[1:5],
  col3 = as.factor(letters[1:5]),
  col4 = as.Date('2020-01-01') + 0:4
)

# Generate the column names and data types report
get_column_info(dataset)
}
