% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn_ts.R
\name{get_fn_ts}
\alias{get_fn_ts}
\title{Get Filename Timestamp}
\usage{
get_fn_ts(time = Sys.time(), tz = "UTC", format = "\%Y-\%B-\%d_\%H-\%M-\%S")
}
\arguments{
\item{time}{A POSIXct object representing the time. Defaults to the current system time (`Sys.time()`).}

\item{tz}{Timezone string}

\item{format}{Format string, R syntax}
}
\value{
A character string representing the cleaned timestamp.
}
\description{
This function generates a timestamp suitable for use in filenames.
The timestamp is cleaned using the `janitor` package to ensure that it
is free from any unwanted characters.
}
\examples{
# Generate a cleaned timestamp for the current time
get_fn_ts()

# Generate a cleaned timestamp for a specific time
get_fn_ts(as.POSIXct("2024-06-15 12:34:56"))
}
