% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dataframe_or_tibble.R
\name{is_dataframe_or_tibble}
\alias{is_dataframe_or_tibble}
\title{Check if an Object is a Data Frame or Tibble}
\usage{
is_dataframe_or_tibble(data)
}
\arguments{
\item{data}{An object to be checked.}
}
\value{
A logical value: `TRUE` if the object is a data frame or tibble, `FALSE` otherwise.
}
\description{
This function checks whether the given object is either a data frame
or a tibble, returning `TRUE` if it is and `FALSE` otherwise.
}
\examples{
# Check if an object is a data frame or tibble
is_dataframe_or_tibble(mtcars)   # TRUE
is_dataframe_or_tibble(tibble::tibble(x = 1:5)) # TRUE
is_dataframe_or_tibble(list(a = 1, b = 2)) # FALSE
}
