% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_core_predicates.R
\name{is_even}
\alias{is_even}
\title{Check if a number is even}
\usage{
is_even(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A logical vector indicating whether each element of `x` is even.
}
\description{
Returns TRUE if the input number is even.
}
