% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_core_predicates.R
\name{is_odd}
\alias{is_odd}
\title{Check if a number is odd}
\usage{
is_odd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A logical vector indicating whether each element of `x` is odd.
}
\description{
Returns TRUE if the input number is odd.
}
