% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_codebook_creator.R
\name{open_codebook_creator}
\alias{open_codebook_creator}
\title{Codebook Generator UI Module}
\usage{
open_codebook_creator(id)
}
\arguments{
\item{id}{A unique module identifier.}
}
\value{
A Shiny UI module for generating a codebook.
}
\description{
This module provides a Shiny UI for uploading a CSV file,
viewing/editing a codebook, and downloading the result.
}
