% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_it}
\alias{read_it}
\title{Read Data File Based on File Extension}
\usage{
read_it(file_path)
}
\arguments{
\item{file_path}{A character string representing the path to the file to be read.}
}
\value{
A data frame or list containing the data from the file.
}
\description{
This function reads a data file based on its extension and returns a data frame or list.
It supports various formats including CSV, JSON, SAS, Parquet, Excel, Feather, and RDS.
}
\examples{
\dontrun{read_it("data.csv")}
\dontrun{read_it("data.json")}
}
