% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_teams_webhook.R
\name{send_teams_webhook_message}
\alias{send_teams_webhook_message}
\title{Send a message to Microsoft Teams using a webhook}
\usage{
send_teams_webhook_message(webhook_url, title, text)
}
\arguments{
\item{webhook_url}{The URL of the Microsoft Teams webhook.}

\item{title}{The title of the message.}

\item{text}{The content of the message.}
}
\value{
A message indicating whether the message was sent successfully or not.
}
\description{
This function sends a message to a specified Microsoft Teams channel using the provided webhook URL.
}
\examples{
webhook_url <- "https://example.webhook.office.com/webhookb2/..."
send_teams_webhook_message(webhook_url, "Test Title", "This is a test message.")
}
