% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataframe.R
\name{stamp_processing_metadata}
\alias{stamp_processing_metadata}
\title{Stamp Processing Metadata}
\usage{
stamp_processing_metadata(df, hash_algo = "md5")
}
\arguments{
\item{df}{A data frame or tibble.}

\item{hash_algo}{A string for the hash algorithm to use (default is "md5").}
}
\value{
A data frame with added `processing_hash` and `processing_timestamp` columns.
}
\description{
Appends hash and timestamp metadata to a data frame.
}
