% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_tibble.R
\name{t_tibble}
\alias{t_tibble}
\title{Transform Data Frame to Tibble in Long Format}
\usage{
t_tibble(.data)
}
\arguments{
\item{.data}{A data frame to be transformed.}
}
\value{
A tibble in long format with key-value pairs.
}
\description{
This function transforms a given data frame into a tibble in a long format,
converting all columns to character type, adding a rowname column, and then
pivoting the data into key-value pairs.
}
\examples{
df <- data.frame(A = 1:3, B = 4:6)
t_tibble(df)

}
