% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{write_any}
\alias{write_any}
\title{Write Data File by Extension}
\usage{
write_any(data, file_path)
}
\arguments{
\item{data}{A data frame or list.}

\item{file_path}{Destination path.}
}
\value{
None.
}
\description{
Writes a data frame to a file based on the file extension.
}
