% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_data_file}
\alias{write_data_file}
\title{Write Data File}
\usage{
write_data_file(data, file_path)
}
\arguments{
\item{data}{A data frame or list to be written to a file.}

\item{file_path}{A character string representing the path to the file where the data should be written.}
}
\description{
This function writes a data frame to a file based on the specified extension.
It supports various file formats including CSV, JSON, Parquet, Excel, Feather, and RDS.
}
\examples{
# Write a data frame to a CSV file
\dontrun{write_data_file(mtcars, "data.csv")}

# Write a data frame to a JSON file
\dontrun{write_data_file(mtcars, "data.json")}

# Write a data frame to a Parquet file
\dontrun{write_data_file(mtcars, "data.parquet")}

# Write a data frame to an Excel file
\dontrun{write_data_file(mtcars, "data.xlsx")}

# Write a data frame to a Feather file
\dontrun{write_data_file(mtcars, "data.feather")}

# Write a data frame to an RDS file
\dontrun{write_data_file(mtcars, "data.rds")}
}
