% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_it}
\alias{write_it}
\title{Write Data to a File Based on Extension}
\usage{
write_it(data, file_path)
}
\arguments{
\item{data}{A data frame or list to be written to a file.}

\item{file_path}{A character string representing the path to the file where the data should be written.}
}
\description{
This function writes a data frame to a file based on the specified extension.
It supports various formats including CSV, JSON, Parquet, Excel, Feather, and RDS.
}
\examples{
\dontrun{write_it(mtcars, "data.csv")}
\dontrun{write_it(mtcars, "data.json")}
}
