% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_date_features.R
\name{append_datetime_features}
\alias{append_datetime_features}
\title{Append Date Features to a Data Frame}
\usage{
append_datetime_features(
  .data,
  dt_var = NA,
  time_features = TRUE,
  time_features_rounded = TRUE
)
}
\arguments{
\item{.data}{A data frame containing the time variable.}

\item{dt_var}{A string specifying the name of the time variable column in the data frame.}

\item{time_features}{A logical for time features}

\item{time_features_rounded}{A logical value indicating whether to append rounded date timestamps. Defaults to TRUE.}
}
\value{
A data frame with appended date and time features.
}
\description{
This function appends various date and time features to a data frame based on a specified time variable.
It can also append rounded date timestamps if requested.
}
\examples{
df <- data.frame(time_stamp = c("2023-06-15 12:34:56", "2023-06-16 13:45:10"))
append_datetime_features(df, dt_var = "time_stamp")

}
