% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blank_pipeline.R
\name{create_blank_pipeline}
\alias{create_blank_pipeline}
\title{Create a Blank Data Analysis Pipeline}
\usage{
create_blank_pipeline(root_path = "", pipeline_folder_name = "demo-pipeline")
}
\arguments{
\item{root_path}{Character. The root path where the pipeline folder should be created. Default is an empty string.}

\item{pipeline_folder_name}{Character. The name of the pipeline folder to be created. Default is "demo-pipeline".}
}
\value{
None. The function creates folders and files in the specified location.
}
\description{
This function creates a new project structure for a data analysis pipeline, including
folders for data, plots, and tidy data, as well as README and TODO files.
}
\examples{
\dontrun{
create_blank_pipeline("~/Desktop", "my_pipeline")
}
}
