% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_report.R
\name{get_package_report}
\alias{get_package_report}
\title{Get Installed Packages Report}
\usage{
get_package_report()
}
\value{
A tibble containing the details of the installed packages.
}
\description{
This function generates a report of the currently installed packages,
returning the information as a cleaned tibble.
}
\examples{
# Generate a report of the installed packages
\dontrun{get_package_report()}

}
