% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_todo_report.R
\name{get_todo_report}
\alias{get_todo_report}
\title{Get TODO Report}
\usage{
get_todo_report()
}
\value{
A tibble containing the details of the TODO comments found in the code.
}
\description{
This function generates a report of all TODO comments detected in the code.
}
\examples{
# Generate a TODO report
\dontrun{get_todo_report()}

}
