% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline_from_config.R
\name{run_pipeline_from_config}
\alias{run_pipeline_from_config}
\title{Run Pipeline from Configuration File}
\usage{
run_pipeline_from_config(config_file = "config.json")
}
\arguments{
\item{config_file}{A character string specifying the path to the configuration file. Defaults to "config.json".}
}
\value{
A tibble containing the results of each pipeline step.
}
\description{
This function reads a configuration file and runs the specified pipeline steps.
}
\examples{
\dontrun{
run_pipeline_from_config("path/to/config.json")
}
}
