% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline_step.R
\name{run_pipeline_step}
\alias{run_pipeline_step}
\title{Run pipeline step with validation}
\usage{
run_pipeline_step(step = NA, pipeline_path = "pipeline/", config = NA)
}
\arguments{
\item{step}{The pipeline step script to run. Should not be NA.}

\item{pipeline_path}{The path to the pipeline scripts, with a trailing slash.}
}
\value{
Integer 1 for success and 0 for failure.
}
\description{
This function runs a specified pipeline step script located in a given pipeline path
after performing a series of validation checks. It logs the start, completion, and
elapsed time of the step execution. Additionally, it can handle different time zones.
}
\examples{
run_pipeline_step(step = "step1.R", pipeline_path = "pipeline/")
}
